version 1.0.2.20 - 1.0.2.41 released
.PMMA end month report redesign
.ADD REJECT OPTION TO IN OUT EDIT PAGE
.ORDER page UI adjust: able to hide order alert.
.give permission store ppl to complete order

+++ FactoryManagementSoftware/FactoryManagementSoftware/BLL/materialUsedBLL.cs
@@ -10,7 +10,7 @@ namespace FactoryManagementSoftware.BLL
     {
         public int no { get; set; }
         public string item_code { get; set; }
-        public int quantity_order { get; set; }
+        public float quantity_order { get; set; }
         public int quantity_order_two { get; set; }
         public int quantity_order_three { get; set; }
         public float item_out { get; set; }
